Colour support removed from form.
Widgets can now be updated to be aware of the form manager.  However, any theme wanting to change the background
of the Form itself will run into problems because of calls to bkgdset within widgets themselves.

In other words, themes that currently rely on BLACK as a background colour for forms will be fine.

Themes that want a different background color will run into difficulty until widgets are updated.